# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BET管理系统的接口文档\n",
        "title": "go-admin API",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/config": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "修改配置",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/config/export": {
            "post": {
                "description": "导出系统配置信息",
                "tags": [
                    "配置"
                ],
                "summary": "导出配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "configId",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/config/{configId}": {
            "delete": {
                "description": "删除数据",
                "tags": [
                    "配置"
                ],
                "summary": "删除配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "configId",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/config/{id}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置编码",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configKey/{configKey}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "configKey",
                        "name": "configKey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configList": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "配置"
                ],
                "summary": "配置列表数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "configKey",
                        "name": "configKey",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "configName",
                        "name": "configName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "configType",
                        "name": "configType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dict/data": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "字典数据"
                ],
                "summary": "修改字典数据",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DictType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "添加配置",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/loginlog": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登录日志"
                ],
                "summary": "修改登录日志",
                "parameters": [
                    {
                        "description": "body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "登录日志"
                ],
                "summary": "添加登录日志",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"添加失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/loginlog/{infoId}": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "通过编码获取登录日志",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "infoId",
                        "name": "infoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除数据",
                "tags": [
                    "登录日志"
                ],
                "summary": "批量删除登录日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "以逗号（,）分割的infoId",
                        "name": "infoId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"删除失败\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/loginloglist": {
            "get": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取JSON",
                "tags": [
                    "登录日志"
                ],
                "summary": "登录日志列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictCode",
                        "name": "dictCode",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "dictType",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": [...]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Config": {
            "type": "object",
            "properties": {
                "configType": {
                    "description": "是否系统内置",
                    "type": "string"
                },
                "content": {
                    "description": "变量字典数据",
                    "type": "string"
                },
                "createBy": {
                    "description": "Params     string `json:\"params\" gorm:\"column:params\"`",
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "extend": {
                    "description": "扩展属性",
                    "type": "string"
                },
                "group": {
                    "description": "变量分组",
                    "type": "string"
                },
                "id": {
                    "description": "编码",
                    "type": "integer"
                },
                "isDel": {
                    "description": "是否删除 0 正常使用，1 已删除",
                    "type": "string"
                },
                "name": {
                    "description": "参数名称 //参数键名ConfigKey string `json:\"configKey\" gorm:\"column:configKey\"`",
                    "type": "string"
                },
                "rule": {
                    "description": "验证规则",
                    "type": "string"
                },
                "tip": {
                    "description": "变量描述 //Remark string `json:\"remark\" gorm:\"column:remark\"` //备注",
                    "type": "string"
                },
                "title": {
                    "description": "变量标题  //参数名称ConfigName string `json:\"Name\" gorm:\"column:name;primary_key\"`",
                    "type": "string"
                },
                "type": {
                    "description": "变量类型:string,text,int,bool,array,datetime,date,file",
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                },
                "value": {
                    "description": "参数变量值 \t//参数键值 //ConfigValue string `json:\"configValue\" gorm:\"column:configValue\"`",
                    "type": "string"
                }
            }
        },
...
        "models.DictType": {
            "type": "object",
            "properties": {
                "createBy": {
                    "description": "创建者",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "type": "string"
                },
                "dictId": {
                    "type": "integer",
                    "example": 1
                },
                "dictName": {
                    "description": "字典名称",
                    "type": "string"
                },
                "dictType": {
                    "description": "字典类型",
                    "type": "string"
                },
                "isDel": {
                    "type": "string"
                },
                "params": {
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "models.LoginLog": {
            "type": "object",
            "properties": {
                "browser": {
                    "description": "浏览器",
                    "type": "string"
                },
                "createBy": {
                    "description": "创建人",
                    "type": "string"
                },
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "dataScope": {
                    "description": "数据",
                    "type": "string"
                },
                "infoId": {
                    "description": "主键",
                    "type": "integer"
                },
                "ipaddr": {
                    "description": "ip地址",
                    "type": "string"
                },
                "isDel": {
                    "description": "是否删除",
                    "type": "string"
                },
                "loginLocation": {
                    "description": "归属地",
                    "type": "string"
                },
                "loginTime": {
                    "description": "登录时间",
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "os": {
                    "description": "系统",
                    "type": "string"
                },
                "params": {
                    "description": "参数",
                    "type": "string"
                },
                "platform": {
                    "description": "固件",
                    "type": "string"
                },
                "remark": {
                    "description": "备注",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "type": "string"
                },
                "updateBy": {
                    "description": "更新者",
                    "type": "string"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "userName": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据集",
                    "type": "object"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}